% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_relaxation.R
\name{genDAGdata}
\alias{genDAGdata}
\title{Title}
\usage{
genDAGdata(n, p, pc)
}
\arguments{
\item{n}{- number of observations}

\item{p}{- number of variables}

\item{pc}{- probability of non zero edges}
}
\value{
n x p data matrix X,  p x p true coefficent matrix B.p
}
\description{
Title
}
\examples{
n = 100
p = 10
pc = 0.2
X = genDAGdata(n, p, pc)
}
