% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_causal_forest.R
\name{rr_test_calibration}
\alias{rr_test_calibration}
\title{ANOVA evaluation of overall detection of treatment effect heterogeneity.}
\usage{
rr_test_calibration(forest, X, Y, W)
}
\arguments{
\item{forest}{The trained forest.}

\item{X}{The covariates used in the causal regression.}

\item{Y}{The outcome (must be a binary numeric vector with no NAs).}

\item{W}{The treatment assignment (must be a binary numeric vector with no NAs).}
}
\value{
P-value for information added by forest predictions.
}
\description{
Test overall TEH detection of the forest. Conducts an ANOVA test to indicate
whether the predicted forest RR coefficients add any information to outcome
predictions, compared to baseline model with only treatment and covariates.
}
